Add allometric equations to a census dataset.

This function adds columns to uniquely identify each row of a census dataset, and each equation from allodb -- so you can look it up on the database.

add_equations(census, equations)

Arguments

census
A ForestGEO-like census dataframe.

equations
An equations dataframe with unique rowids.

Value

A dataframe with all columns from census plus additional columns rowid, eqn equation_id.

See also

Examples

census <- dplyr::sample_n(fgeo.biomass::scbi_tree1, 30)
species <- fgeo.biomass::scbi_species

single_best <- census %>%
  add_species(species, site = "scbi") %>%
  allo_find() %>%
  tidyr::unnest() %>%
  fixme_drop_duplicated_rowid()

#> Adding `site`.

#> Overwriting `sp`; it now stores Latin species names.

#> Adding `rowid`.

#> Warning: Can't find equations matching these species (inserting 2 missing values):
#> quercus prinus, unidentified unk

#> Joining, by = sp, site.

#> Warning: Dropping 5 rows with duplicated `rowid` values.

add_equations(census, single_best)

#> Joining, by = "rowid"

#> # A tibble: 30 x 23
#>    rowid treeID stemID tag   StemTag sp    quadrat    gx     gy DBHID CensusID
#>    <int>  <int>  <int> <chr> <chr>   <chr> <chr>   <dbl>  <dbl> <int>    <int>
#>  1     1   3253   3253 22229 1       astr  0223     37.7 443     6204        1
#>  2     2  33609     NA 1435… <NA>    astr  1415    273.  291.      NA       NA
#>  3     3  24200  24200 1702… 1       libe  1701    340     6.40 31811        1
#>  4     4   6333   6333 50450 1       caca  0508     95.7 148.    9963        1
#>  5     5    299    299 1008… 1       libe  1005    194.   91.7    989        1
#>  6     6  18786  18786 1324… 1       unk   1332    242.  629.   25246        1
#>  7     7  20049  20049 1420… 1       astr  1419    276.  364.   26713        1
#>  8     8  11671  11671 92527 1       litu  0930    177   599.   16406        1
#>  9     9  29517  29517 2004… 1       libe  2003    388.   44.7  38264        1
#> 10    10  31113  31113 2027… 1       ceoc  2027    399.  538.   40100        1
#> # … with 20 more rows, and 12 more variables: dbh <dbl>, pom <chr>, hom <dbl>,
#> #   ExactDate <chr>, DFstatus <chr>, codes <chr>, nostems <dbl>, date <dbl>,
#> #   status <chr>, agb <dbl>, eqn <chr>, equation_id <chr>