Add biomass using allometric equations for tropical trees.

This function wraps a number of functions from the BIOMASS package, which you may see for more options and details. It uses pantropical models from Chave et al. 2014 to estimate the above-ground biomass of tropical trees.

add_tropical_biomass(data, species, region = "Pantropical",
  latitude = NULL, longitude = NULL,
  dbh_unit = guess_dbh_unit(data$dbh))

Arguments

data
A ForestGEO-like census-dataframe.

species
A ForestGEO-like species-dataframe.

region
Area of your dataset to estimate tree height thanks to Weibull-H region-, continent-specific and pantropical models proposed by Feldpausch et al. (2012). To be chosen between:

Africa: Africa

CAfrica: Central Africa

EAfrica: Eastern Africa

WAfrica: Western Africa

SAmerica: Southern America

BrazilianShield: Brazilian Shield

ECAmazonia: East-Central Amazonia

GuianaShield: Guiana Shield

WAmazonia: Western Amazonia

SEAsia: South-Eastern Asia

NAustralia: Northern Australia

Pantropical: Pantropical

latitude, longitude
A number giving coordinates, e.g. latitude = 9.004080, longitude = -79.525635. It can also be a vector of such numbers, with as many elements as the number of rows of data.

dbh_unit
Character string giving the unit of dbh values, e.g. "mm" (see valid_units()).

Value

A modified version of the data dataframe, with additional columns giving taxonomic, wood density (in g/cm^3), and biomass (in kg) information.

References

Chave et al. (2014) Improved allometric models to estimate the aboveground biomass of tropical trees, Global Change Biology, 20 (10), 3177-3190

See also

Examples

library(dplyr)
library(fgeo.biomass)

data <- fgeo.biomass::scbi_stem_tiny_tree
species <- fgeo.biomass::scbi_species

add_tropical_biomass(data, species, region = "pantropical")

#> ✔ Guessing dbh in [mm].
#> ℹ You may provide the dbh unit manually via the argument`dbh_unit`.
#> ℹ Wood density given in [g/cm^3].
#> ✔ Using 'pantropical' `region`.
#> ℹ Biomass is given in [kg].
#> ✔ Adding new columns:
#>   family, genus, species, wd_level, wd_mean, wd_sd, biomass

#> # A tibble: 11 x 27
#>    treeID stemID    tag StemTag sp    quadrat    gx    gy DBHID CensusID   dbh
#>     <dbl>  <dbl>  <dbl>   <dbl> <chr> <chr>   <dbl> <dbl> <dbl>    <dbl> <dbl>
#>  1   3175   3175  22146       1 caca  0221     29    404.  6103        1 10.7 
#>  2   3175  32574  22146       2 caca  0221     29    404.  6104        1  3.14
#>  3   3996   3996  30363       1 litu  0307     50.3  139.  7096        1 75.5 
#>  4   3996  32682  30363       2 litu  0307     50.3  139.  7097        1 12.7 
#>  5   5570   5570  42082       1 caca  0419     71.3  363.  9047        1  9.83
#>  6   5570  32894  42082       2 caca  0419     71.3  363.  9048        1  4.93
#>  7   5570  37222  42082       3 caca  0419     71.3  363.  9049        1  4.43
#>  8  10060  10060  82218       1 litu  0822    159.   436  14579        1 61.4 
#>  9  10060  33449  82218       2 litu  0822    159.   436  14580        1 22.2 
#> 10  26760  26760 180969       1 ulru  1812    350.   227. 34958        1 34.3 
#> 11  26760  35964 180969       2 ulru  1812    350.   227. 34959        1  5.03
#> # … with 16 more variables: pom <dbl>, hom <dbl>, ExactDate <date>,
#> #   DFstatus <chr>, codes <chr>, countPOM <dbl>, date <dbl>, status <chr>,
#> #   agb <dbl>, family <chr>, genus <chr>, species <chr>, wd_level <chr>,
#> #   wd_mean <dbl>, wd_sd <dbl>, biomass <dbl>

# Not running to reduce build check-time

# NOT RUN {
data %>%
  add_tropical_biomass(species, latitude = -34, longitude = -58) %>%
  select(biomass, everything())
# }