Add wood density.

This function wraps some features of BIOMASS::getWoodDensity(), which you may see for more details and options. It estimates the wood density of the trees from their taxonomy or from their congeners using the global wood density database (Chave et al. 2009, Zanne et al. 2009) or any additional dataset. The resulting values of wood density can either be attributed to an individual at the level of a species, genus, or family.

add_wood_density(data, species, family = TRUE, region = "World")

Arguments

data
A ForestGEO-like census-dataframe.

species
A ForestGEO-like species-dataframe.

family
(optional) Vector of families. If set, the missing wood densities at the genus level will be attributed at family level if available.

region
Region (or vector of region) of interest of your sample. By default, Region is set to 'World', but you can restrict the WD estimates to a single region :

AfricaExtraTrop: Africa (extra tropical)

AfricaTrop: Africa (tropical)

Australia: Australia

AustraliaTrop: Australia (tropical)

CentralAmericaTrop: Central America (tropical)

China: China

Europe: Europe

India: India

Madagascar: Madagascar

Mexico: Mexico

NorthAmerica: North America

Oceania: Oceania

SouthEastAsia: South-East Asia

SouthEastAsiaTrop: South-East Asia (tropical)

SouthAmericaExtraTrop: South America (extra tropical)

SouthAmericaTrop: South America (tropical)

World: World

Value

A dataframe as that passed to data, but with additional columns giving taxonomic information, and the following wood density information (oven dry mass/fresh volume in g/cm^3):

wd_mean: Mean wood density.

wd_sd: Standard deviation of the wood density that can be used in error propagation (see sd_10 and AGBmonteCarlo()).

wd_level: Level at which wood density has been calculated. Can be species, genus, family, dataset (mean of the entire dataset) or, if stand is set, the name of the stand (mean of the current stand).

Details

This function assigns to each taxon a species- or genus- level average if at least one wood density value at the genus level is available for that taxon in the reference database. If not, the mean wood density of the family (if family = TRUE).

As an estimate of the error associated with the wood density estimate, this function also provides the mean standard deviation value at the appropriate taxonomic level.

References

Rejou-Mechain M, Tanguy A, Piponiot C, Chave J, Herault B (2017). “BIOMASS : an R package for estimating above-ground biomass and its uncertainty in tropical forests.” Methods in Ecology and Evolution, 8(9). ISSN 2041210X, doi: 10.1111/2041-210X.12753 (URL: http://doi.org/10.1111/2041-210X.12753), <URL: http://doi.wiley.com/10.1111/2041-210X.12753>.

See also

Examples

tree <- fgeo.biomass::scbi_stem_tiny_tree
species <- fgeo.biomass::scbi_species
add_wood_density(tree, species)

#> ℹ Wood density given in [g/cm^3].

#> # A tibble: 11 x 26
#>    treeID stemID    tag StemTag sp    quadrat    gx    gy DBHID CensusID   dbh
#>     <dbl>  <dbl>  <dbl>   <dbl> <chr> <chr>   <dbl> <dbl> <dbl>    <dbl> <dbl>
#>  1   3175   3175  22146       1 caca  0221     29    404.  6103        1 107  
#>  2   3175  32574  22146       2 caca  0221     29    404.  6104        1  31.4
#>  3   3996   3996  30363       1 litu  0307     50.3  139.  7096        1 755. 
#>  4   3996  32682  30363       2 litu  0307     50.3  139.  7097        1 127. 
#>  5   5570   5570  42082       1 caca  0419     71.3  363.  9047        1  98.3
#>  6   5570  32894  42082       2 caca  0419     71.3  363.  9048        1  49.3
#>  7   5570  37222  42082       3 caca  0419     71.3  363.  9049        1  44.3
#>  8  10060  10060  82218       1 litu  0822    159.   436  14579        1 614. 
#>  9  10060  33449  82218       2 litu  0822    159.   436  14580        1 222. 
#> 10  26760  26760 180969       1 ulru  1812    350.   227. 34958        1 343. 
#> 11  26760  35964 180969       2 ulru  1812    350.   227. 34959        1  50.3
#> # … with 15 more variables: pom <dbl>, hom <dbl>, ExactDate <date>,
#> #   DFstatus <chr>, codes <chr>, countPOM <dbl>, date <dbl>, status <chr>,
#> #   agb <dbl>, family <chr>, genus <chr>, species <chr>, wd_level <chr>,
#> #   wd_mean <dbl>, wd_sd <dbl>