Evaluate equations, giving a biomass result per row.

Evaluate equations, giving a biomass result per row.

allo_evaluate(data, biomass_unit = "kg")

Arguments

data
A dataframe as those created with allo_find().

biomass_unit
Character string giving the output unit e.g. "kg".

Value

A dataframe with a single row by each value of rowid.

See also

Examples

library(dplyr)

best <- fgeo.biomass::scbi_tree1 %>%
  # Pick few rows for a quick example
  sample_n(500) %>%
  add_species(fgeo.biomass::scbi_species, "scbi") %>%
  allo_find()

#> Adding `site`.

#> Overwriting `sp`; it now stores Latin species names.

#> Adding `rowid`.

#> Assuming `dbh` in [mm] (required to find dbh-specific equations).

#> * Searching equations according to site and species.

#> Warning: Can't find equations matching these species:
#> carya sp, crataegus sp, quercus prinus, ulmus sp, unidentified unk

#> * Refining equations according to dbh.

#> Warning:   Can't find equations for 314 rows (inserting `NA`).

allo_evaluate(best)

#> Assuming `dbh` unit in [mm].

#> Converting `dbh` based on `dbh_unit`.

#> `biomass` values are given in [kg].

#> Warning: Can't convert all units (inserting 314 missing values):
#> the 'to' argument is not an acceptable unit.

#> Warning: Can't convert all units (inserting 314 missing values):
#> the 'from' argument is not an acceptable unit.

#> Warning: Can't evaluate all equations (inserting 4 missing values):
#> object 'dba' not found

#> Warning: `biomass` may be invalid. This is still work in progress.

#> # A tibble: 500 x 2
#>    rowid biomass
#>    <int>   <dbl>
#>  1     1   NA   
#>  2     2   NA   
#>  3     3   NA   
#>  4     4   18.5 
#>  5     5   NA   
#>  6     6   NA   
#>  7     7    1.16
#>  8     8 2222.  
#>  9     9   NA   
#> 10    10    8.75
#> # … with 490 more rows

allo_evaluate(best, biomass_unit = "Mg")

#> Assuming `dbh` unit in [mm].

#> Converting `dbh` based on `dbh_unit`.

#> `biomass` values are given in [Mg].

#> Warning: Can't convert all units (inserting 314 missing values):
#> the 'to' argument is not an acceptable unit.

#> Warning: Can't convert all units (inserting 314 missing values):
#> the 'from' argument is not an acceptable unit.

#> Warning: Can't evaluate all equations (inserting 4 missing values):
#> object 'dba' not found

#> Warning: `biomass` may be invalid. This is still work in progress.

#> # A tibble: 500 x 2
#>    rowid  biomass
#>    <int>    <dbl>
#>  1     1 NA      
#>  2     2 NA      
#>  3     3 NA      
#>  4     4  0.0185 
#>  5     5 NA      
#>  6     6 NA      
#>  7     7  0.00116
#>  8     8  2.22   
#>  9     9 NA      
#> 10    10  0.00875
#> # … with 490 more rows