Find allometric equations in allodb or in a custom equations-table.

Find allometric equations in allodb or in a custom equations-table.

allo_find(dbh_species, dbh_unit = "mm", custom_eqn = NULL)

Arguments

dbh_species
A dataframe as those created with add_species().

dbh_unit
Character string giving the unit of the expected input dbh.

custom_eqn
A dataframe of class "eqn".

Value

A nested dataframe with each row containing the data of an equation type.

See also

Examples

census <- dplyr::sample_n(fgeo.biomass::scbi_tree1, 30)
species <- fgeo.biomass::scbi_species
census_species <- add_species(
  census, species,
  site = "scbi"
)

#> Adding `site`.

#> Overwriting `sp`; it now stores Latin species names.

#> Adding `rowid`.

allo_find(census_species)

#> Assuming `dbh` data in [mm].

#> Joining, by = c("sp", "site")

#> Converting `dbh` based on `dbh_unit`.

#> Warning: Can't convert all units (inserting 7 missing values):
#> the 'to' argument is not an acceptable unit.

#> # A tibble: 30 x 29
#>    rowid treeID stemID tag   StemTag sp    quadrat    gx     gy DBHID CensusID
#>    <int>  <int>  <int> <chr> <chr>   <chr> <chr>   <dbl>  <dbl> <int>    <int>
#>  1     1   7471   7471 60486 1       fagu… 0616    106   316.   11367        1
#>  2     2  13149  13149 1022… 1       cary… 1021    191.  409.   18241        1
#>  3     3  32913     NA 1333… <NA>    lind… 1310    258.  181.      NA       NA
#>  4     4  30160  30160 2011… 1       lind… 2010    385   198.   39052        1
#>  5     5  21419  21419 1508… 1       lind… 1509    297.  180.   28380        1
#>  6     6  38731     NA 33008 <NA>    lind… 0301     46.5   4.20    NA       NA
#>  7     7   4191   4191 30568 1       quer… 0316     49.2 303.    7317        1
#>  8     8  29689  29689 2006… 1       ulmu… 2006    400   120    38457        1
#>  9     9  34133     NA 1533… <NA>    ulmu… 1509    299.  179.      NA       NA
#> 10    10  25926  25926 1800… 1       ulmu… 1801    345.    7.80 33900        1
#> # … with 20 more rows, and 18 more variables: dbh <dbl>, pom <chr>, hom <dbl>,
#> #   ExactDate <chr>, DFstatus <chr>, codes <chr>, nostems <dbl>, date <dbl>,
#> #   status <chr>, agb <dbl>, site <chr>, equation_id <chr>, eqn <chr>,
#> #   eqn_source <chr>, eqn_type <chr>, anatomic_relevance <chr>, dbh_unit <chr>,
#> #   bms_unit <chr>

allo_find(census_species, dbh_unit = "cm")

#> Assuming `dbh` data in [cm].

#> Joining, by = c("sp", "site")

#> Converting `dbh` based on `dbh_unit`.

#> Warning: Can't convert all units (inserting 7 missing values):
#> the 'to' argument is not an acceptable unit.

#> # A tibble: 30 x 29
#>    rowid treeID stemID tag   StemTag sp    quadrat    gx     gy DBHID CensusID
#>    <int>  <int>  <int> <chr> <chr>   <chr> <chr>   <dbl>  <dbl> <int>    <int>
#>  1     1   7471   7471 60486 1       fagu… 0616    106   316.   11367        1
#>  2     2  13149  13149 1022… 1       cary… 1021    191.  409.   18241        1
#>  3     3  32913     NA 1333… <NA>    lind… 1310    258.  181.      NA       NA
#>  4     4  30160  30160 2011… 1       lind… 2010    385   198.   39052        1
#>  5     5  21419  21419 1508… 1       lind… 1509    297.  180.   28380        1
#>  6     6  38731     NA 33008 <NA>    lind… 0301     46.5   4.20    NA       NA
#>  7     7   4191   4191 30568 1       quer… 0316     49.2 303.    7317        1
#>  8     8  29689  29689 2006… 1       ulmu… 2006    400   120    38457        1
#>  9     9  34133     NA 1533… <NA>    ulmu… 1509    299.  179.      NA       NA
#> 10    10  25926  25926 1800… 1       ulmu… 1801    345.    7.80 33900        1
#> # … with 20 more rows, and 18 more variables: dbh <dbl>, pom <chr>, hom <dbl>,
#> #   ExactDate <chr>, DFstatus <chr>, codes <chr>, nostems <dbl>, date <dbl>,
#> #   status <chr>, agb <dbl>, site <chr>, equation_id <chr>, eqn <chr>,
#> #   eqn_source <chr>, eqn_type <chr>, anatomic_relevance <chr>, dbh_unit <chr>,
#> #   bms_unit <chr>

# PROVIDE CUSTOM EQUAITONS ----------------------------------------------
# Checks that the structure of your data isn't terriby wrong
# BAD
try(as_eqn("really bad data"))

#> Error in validate_eqn(data) : is.data.frame(data) is not TRUE

try(as_eqn(data.frame(1)))

#> Error : Ensure your data set has these variables:
#> equation_id, site, sp, eqn, eqn_type, anatomic_relevance, dbh_unit, bms_unit

# GOOD
your_equations <- tibble::tibble(
  equation_id = c("000001"),
  site = c("scbi"),
  sp = c("tilia americana"),
  # Watning: Fake!
  eqn = c("exp(-2.48 + 2.4835 * log(dbh))"),
  eqn_type = c("mixed_hardwood"),
  anatomic_relevance = c("total aboveground biomass"),
 dbh_unit = "cm",
 bms_unit = "g"
)

class(as_eqn(your_equations))

#> [1] "eqn"        "tbl_df"     "tbl"        "data.frame"

allo_find(census_species, custom_eqn = as_eqn(your_equations))

#> Assuming `dbh` data in [mm].

#> Joining, by = c("sp", "site")

#> Converting `dbh` based on `dbh_unit`.

#> Warning: Can't convert all units (inserting 30 missing values):
#> the 'to' argument is not an acceptable unit.

#> # A tibble: 30 x 29
#>    rowid treeID stemID tag   StemTag sp    quadrat    gx     gy DBHID CensusID
#>    <int>  <int>  <int> <chr> <chr>   <chr> <chr>   <dbl>  <dbl> <int>    <int>
#>  1     1   7471   7471 60486 1       fagu… 0616    106   316.   11367        1
#>  2     2  13149  13149 1022… 1       cary… 1021    191.  409.   18241        1
#>  3     3  32913     NA 1333… <NA>    lind… 1310    258.  181.      NA       NA
#>  4     4  30160  30160 2011… 1       lind… 2010    385   198.   39052        1
#>  5     5  21419  21419 1508… 1       lind… 1509    297.  180.   28380        1
#>  6     6  38731     NA 33008 <NA>    lind… 0301     46.5   4.20    NA       NA
#>  7     7   4191   4191 30568 1       quer… 0316     49.2 303.    7317        1
#>  8     8  29689  29689 2006… 1       ulmu… 2006    400   120    38457        1
#>  9     9  34133     NA 1533… <NA>    ulmu… 1509    299.  179.      NA       NA
#> 10    10  25926  25926 1800… 1       ulmu… 1801    345.    7.80 33900        1
#> # … with 20 more rows, and 18 more variables: dbh <dbl>, pom <chr>, hom <dbl>,
#> #   ExactDate <chr>, DFstatus <chr>, codes <chr>, nostems <dbl>, date <dbl>,
#> #   status <chr>, agb <dbl>, site <chr>, equation_id <chr>, eqn <chr>,
#> #   eqn_source <chr>, eqn_type <chr>, anatomic_relevance <chr>, dbh_unit <chr>,
#> #   bms_unit <chr>

census_species %>%
allo_find(custom_eqn = as_eqn(your_equations))

#> # A tibble: 30 x 29
#>    rowid treeID stemID tag   StemTag sp    quadrat    gx     gy DBHID CensusID
#>    <int>  <int>  <int> <chr> <chr>   <chr> <chr>   <dbl>  <dbl> <int>    <int>
#>  1     1   7471   7471 60486 1       fagu… 0616    106   316.   11367        1
#>  2     2  13149  13149 1022… 1       cary… 1021    191.  409.   18241        1
#>  3     3  32913     NA 1333… <NA>    lind… 1310    258.  181.      NA       NA
#>  4     4  30160  30160 2011… 1       lind… 2010    385   198.   39052        1
#>  5     5  21419  21419 1508… 1       lind… 1509    297.  180.   28380        1
#>  6     6  38731     NA 33008 <NA>    lind… 0301     46.5   4.20    NA       NA
#>  7     7   4191   4191 30568 1       quer… 0316     49.2 303.    7317        1
#>  8     8  29689  29689 2006… 1       ulmu… 2006    400   120    38457        1
#>  9     9  34133     NA 1533… <NA>    ulmu… 1509    299.  179.      NA       NA
#> 10    10  25926  25926 1800… 1       ulmu… 1801    345.    7.80 33900        1
#> # … with 20 more rows, and 18 more variables: dbh <dbl>, pom <chr>, hom <dbl>,
#> #   ExactDate <chr>, DFstatus <chr>, codes <chr>, nostems <dbl>, date <dbl>,
#> #   status <chr>, agb <dbl>, site <chr>, equation_id <chr>, eqn <chr>,
#> #   eqn_source <chr>, eqn_type <chr>, anatomic_relevance <chr>, dbh_unit <chr>,
#> #   bms_unit <chr>