Find the best equations given a priority order.

This function orders the output of allo_find() by equation type, then reduces it in order, by overwriting equations of each type with the equations of a higher-priority type.

allo_order(.data, order = NULL)

Arguments

.data
List of dataframes. Each should have a rowid column giving the index of each row. Otherwise, rowid will be added with a warning.

order
String giving the name or index of the list elements in the order they should be row-bind, with elements on the left winning over elements on the right.

Value

A dataframe.

See also

Examples

allodb::scbi_tree1 %>%
  add_species(allodb::scbi_species, "scbi") %>%
  allo_find() %>%
  allo_order()

#> # A tibble: 30,229 x 8
#>    eqn_type rowid site  sp           dbh equation_id eqn             eqn_source
#>    <chr>    <int> <chr> <chr>      <dbl> <chr>       <chr>           <chr>     
#>  1 family       1 scbi  lindera b~  27.9 f08fff      exp(-2.2118 + ~ default   
#>  2 family       2 scbi  lindera b~  23.7 f08fff      exp(-2.2118 + ~ default   
#>  3 family       3 scbi  lindera b~  22.2 f08fff      exp(-2.2118 + ~ default   
#>  4 family       8 scbi  lindera b~  51.4 f08fff      exp(-2.2118 + ~ default   
#>  5 family      13 scbi  lindera b~  15.4 f08fff      exp(-2.2118 + ~ default   
#>  6 family      14 scbi  lindera b~  14.8 f08fff      exp(-2.2118 + ~ default   
#>  7 family      15 scbi  lindera b~  15.5 f08fff      exp(-2.2118 + ~ default   
#>  8 family      16 scbi  lindera b~  17.4 f08fff      exp(-2.2118 + ~ default   
#>  9 family      17 scbi  lindera b~  68.2 f08fff      exp(-2.2118 + ~ default   
#> 10 family      18 scbi  lindera b~  19.3 f08fff      exp(-2.2118 + ~ default   
#> # ... with 30,219 more rows