A safe version of measurements::conv_unit().

Unlike measurements::conv_unit() this function doesn't fail if it can't convert a unit -- instead it outputs NA.

convert_units(x, from, to)

Arguments

x
a numeric vector giving the measurement value in its original units.

from
the unit in which the measurement was made.

to
the unit to which the measurement is to be converted.

Value

As in measurements::conv_unit() if possible, or NA.

See also

Examples

convert_units(c(1, NA), from = "cm", to = "m")

#> [1] 0.01   NA

# Same
measurements::conv_unit(c(1, NA), from = "cm", to = "m")

#> [1] 0.01   NA

# Passes
convert_units(c(1, 10), from = "cm", to = c("m", "bad"))

#> [1] 0.01   NA

# Errs
try(measurements::conv_unit(c(1, 10), from = "cm", to = c("m", "bad")))

#> Error in measurements::conv_unit(c(1, 10), from = "cm", to = c("m", "bad")) : 
#>   the 'to' argument is not an acceptable unit.