Exclude equations that can't be evaluated.

This function tries to evaluate allodb equations and excludes the ones that err.

fixme_exclude_failing_equations(data)

Arguments

data
An equations dataframe, e.g. allodb::master().

Value

A dataframe with all equations that can be successfully evaluated.

See also

Examples

fixme_exclude_failing_equations(allodb::master())

#> # A tibble: 689 x 43
#>    ref_id equation_allome… equation_id equation_form dependent_varia…
#>    <chr>  <chr>            <chr>       <chr>         <chr>           
#>  1 jenki… 10^(1.1891+1.41… 2060ea      10^(a+b*(log… Total abovegrou…
#>  2 jenki… 10^(1.1891+1.41… 2060ea      10^(a+b*(log… Total abovegrou…
#>  3 jenki… 10^(1.2315+1.63… a4d879      10^(a+b*(log… Total abovegrou…
#>  4 jenki… 10^(1.2315+1.63… a4d879      10^(a+b*(log… Total abovegrou…
#>  5 jenki… exp(7.217+1.514… c59e03      exp(a+b*log(… Stem biomass (w…
#>  6 jenki… exp(7.217+1.514… c59e03      exp(a+b*log(… Stem biomass (w…
#>  7 jenki… exp(7.217+1.514… c59e03      exp(a+b*log(… Stem biomass (w…
#>  8 jenki… exp(7.217+1.514… c59e03      exp(a+b*log(… Stem biomass (w…
#>  9 jenki… exp(7.217+1.514… c59e03      exp(a+b*log(… Stem biomass (w…
#> 10 jenki… exp(7.217+1.514… c59e03      exp(a+b*log(… Stem biomass (w…
#> # … with 679 more rows, and 38 more variables: independent_variable <chr>,
#> #   allometry_specificity <chr>, geographic_area <chr>, dbh_min_cm <chr>,
#> #   dbh_max_cm <chr>, sample_size <chr>, dbh_units_original <chr>,
#> #   biomass_units_original <chr>, allometry_development_method <chr>,
#> #   regression_model <chr>, other_equations_tested <chr>, log_biomass <chr>,
#> #   bias_corrected <chr>, bias_correction_factor <chr>,
#> #   notes_fitting_model <chr>, original_data_availability <chr>, warning <chr>,
#> #   site <chr>, family <chr>, species <chr>, species_code <chr>,
#> #   life_form <chr>, equation_group <chr>, equation_taxa <chr>,
#> #   notes_on_species <chr>, wsg_id <chr>, wsg_specificity <chr>, id <chr>,
#> #   Site <chr>, lat <chr>, long <chr>, UTM_Zone <chr>, UTM_X <chr>,
#> #   UTM_Y <chr>, intertropical <chr>, size.ha <chr>, E <chr>,
#> #   wsg.site.name <chr>