Handle multiple rowids: find duplicates and pick a single row.

Handle multiple rowids: find duplicates and pick a single row.

fixme_find_duplicated_rowid(data)

fixme_drop_duplicated_rowid(data)

Arguments

data
A dataframe as those created with allo_find().

Value

fixme_find_duplicated_rowid() returns a dataframe with the rows for which duplicated rowid values are found.

fixme_drop_duplicated_rowid() returns a dataframe with a single row per rowid value.

See also

Examples

best <- fgeo.biomass::scbi_tree1 %>%
  add_species(fgeo.biomass::scbi_species, "scbi") %>%
  allo_find()

#> Adding `site`.

#> Overwriting `sp`; it now stores Latin species names.

#> Adding `rowid`.

best %>%
  fixme_find_duplicated_rowid()

#> # A tibble: 1,823 x 30
#>    rowid treeID stemID tag   StemTag sp    quadrat    gx    gy DBHID CensusID
#>    <int>  <int>  <int> <chr> <chr>   <chr> <chr>   <dbl> <dbl> <int>    <int>
#>  1   106    106    106 40452 1       amel… 0404     64.4  61.5   379        1
#>  2   106    106    106 40452 1       amel… 0404     64.4  61.5   379        1
#>  3   106    106    106 40452 1       amel… 0404     64.4  61.5   379        1
#>  4   125    125    125 41047 1       amel… 0403     75.4  57.8   436        1
#>  5   125    125    125 41047 1       amel… 0403     75.4  57.8   436        1
#>  6   125    125    125 41047 1       amel… 0403     75.4  57.8   436        1
#>  7   131    131    131 50171 1       amel… 0503     85.3  47     462        1
#>  8   131    131    131 50171 1       amel… 0503     85.3  47     462        1
#>  9   131    131    131 50171 1       amel… 0503     85.3  47     462        1
#> 10   132    132    132 50199 1       amel… 0504     87.6  70.2   465        1
#> # … with 1,813 more rows, and 19 more variables: dbh <dbl>, pom <chr>,
#> #   hom <dbl>, ExactDate <chr>, DFstatus <chr>, codes <chr>, nostems <dbl>,
#> #   date <dbl>, status <chr>, agb <dbl>, site <chr>, equation_id <chr>,
#> #   eqn <chr>, eqn_source <chr>, eqn_type <chr>, anatomic_relevance <chr>,
#> #   dbh_unit <chr>, bms_unit <chr>, n <int>

best %>%
  fixme_drop_duplicated_rowid()

#> Warning: Dropping 788 rows with duplicated `rowid` values.

#> # A tibble: 39,495 x 29
#>    rowid treeID stemID tag   StemTag sp    quadrat    gx    gy DBHID CensusID
#>    <int>  <int>  <int> <chr> <chr>   <chr> <chr>   <dbl> <dbl> <int>    <int>
#>  1     1      1      1 10079 1       lind… 0104     3.70  73       1        1
#>  2     2      2      2 10168 1       lind… 0103    17.3   58.9     3        1
#>  3     3      3      3 10567 1       lind… 0110     9    197.      5        1
#>  4     4      4      4 12165 1       nyss… 0122    14.2  428.      7        1
#>  5     5      5      5 12190 1       hama… 0122     9.40 436.      9        1
#>  6     6      6      6 12192 1       hama… 0122     1.30 434      13        1
#>  7     7      7      7 12212 1       unid… 0123    17.8  447.     15        1
#>  8     8      8      8 12261 1       lind… 0125    18    484.     17        1
#>  9     9      9      9 12456 1       vibu… 0130    18    598.     19        1
#> 10    10     10     10 12551 1       asim… 0132     5.60 628.     22        1
#> # … with 39,485 more rows, and 18 more variables: dbh <dbl>, pom <chr>,
#> #   hom <dbl>, ExactDate <chr>, DFstatus <chr>, codes <chr>, nostems <dbl>,
#> #   date <dbl>, status <chr>, agb <dbl>, site <chr>, equation_id <chr>,
#> #   eqn <chr>, eqn_source <chr>, eqn_type <chr>, anatomic_relevance <chr>,
#> #   dbh_unit <chr>, bms_unit <chr>

# Should return 0-rows
best %>%
  fixme_drop_duplicated_rowid() %>%
  fixme_find_duplicated_rowid()

#> Warning: Dropping 788 rows with duplicated `rowid` values.

#> # A tibble: 0 x 30
#> # … with 30 variables: rowid <int>, treeID <int>, stemID <int>, tag <chr>,
#> #   StemTag <chr>, sp <chr>, quadrat <chr>, gx <dbl>, gy <dbl>, DBHID <int>,
#> #   CensusID <int>, dbh <dbl>, pom <chr>, hom <dbl>, ExactDate <chr>,
#> #   DFstatus <chr>, codes <chr>, nostems <dbl>, date <dbl>, status <chr>,
#> #   agb <dbl>, site <chr>, equation_id <chr>, eqn <chr>, eqn_source <chr>,
#> #   eqn_type <chr>, anatomic_relevance <chr>, dbh_unit <chr>, bms_unit <chr>,
#> #   n <int>