Propagate errors in above-ground biomass (AGB).

This functions is warps BIOMASS::AGBmonteCarlo(), which you may use directly for more options.

propagate_errors(data, n = 1000, dbh_sd = NULL, height_model = NULL)

Arguments

data
The output of add_tropical_biomass().

n
Number of iterations. Cannot be smaller than 50 or larger than 1000. By default n = 1000

dbh_sd
This variable can take three kind of values, indicating how to propagate the errors on diameter measurements: a single numerical value or a vector of the same size as D, both representing the standard deviation associated with the diameter measurements or "chave2004" (an important error on 5 percent of the measures, a smaller error on 95 percent of the trees).

height_model
Model used to estimate tree height from tree diameter (output from model_height(), see example).

Value

A list with the following elements:

meanAGB: Mean stand AGB value following the error propagation.

medAGB: Median stand AGB value following the error propagation.

sdAGB: Standard deviation of the stand AGB value following the error propagation.

credibilityAGB: Credibility interval at 95 following the error propagation.

AGB_simu: Matrix with the AGB of the trees (rows) times the n iterations (columns).

Details

See Rejou-Mechain et al. (2017) for all details on the error propagation procedure.

References

Chave, J. et al. (2004). Error propagation and scaling for tropical forest biomass estimates. Philosophical Transactions of the Royal Society B: Biological Sciences, 359(1443), 409-420.

Rejou-Mechain et al. (2017). BIOMASS: An R Package for estimating above-ground biomass and its uncertainty in tropical forests. Methods in Ecology and Evolution, 8 (9), 1163-1167.

Examples

library(dplyr)

data <- fgeo.biomass::scbi_tree1 %>%
  slice(1:100)
species <- fgeo.biomass::scbi_species

# Using `region` (default)
biomass <- add_tropical_biomass(data, species)

#> ✔ Guessing dbh in [mm].
#> ℹ You may provide the dbh unit manually via the argument`dbh_unit`.
#> ℹ Wood density given in [g/cm^3].
#> ✔ Using 'Pantropical' `region`.
#> ℹ Biomass is given in [kg].
#> ✔ Adding new columns:
#>   family, genus, species, wd_level, wd_mean, wd_sd, biomass

model <- model_height(biomass)

#> ℹ Using `method` log1 (other methods: log2, weibull, michaelis).

str(
  propagate_errors(biomass, n = 50, height_model = model)
)

#> ✔ Propagating errors on measurements of wood density.
#> ✔ Propagating errors on measurements of height.
#> List of 5
#>  $ meanAGB       : num 11.2
#>  $ medAGB        : num 10.8
#>  $ sdAGB         : num 2.41
#>  $ credibilityAGB: Named num [1:2] 6.97 17.13
#>   ..- attr(*, "names")= chr [1:2] "2.5%" "97.5%"
#>  $ AGB_simu      : num [1:100, 1:50] 0.000658 0.000462 0.000284 0.087518 0.027327 ...

# Using `latitude` and `longitude`
biomass <- add_tropical_biomass(
  data = data,
  species = species,
  latitude = 4,
  longitude = -52
)

#> ✔ Guessing dbh in [mm].
#> ℹ You may provide the dbh unit manually via the argument`dbh_unit`.
#> ℹ Wood density given in [g/cm^3].
#> ✔ Using `latitude` and `longitude` (ignoring `region`).
#> ℹ Biomass is given in [kg].
#> ✔ Adding new columns:
#>   family, genus, species, wd_level, wd_mean, wd_sd, latitude, longitude, biomass

model <- model_height(biomass)

#> ℹ Using `method` log1 (other methods: log2, weibull, michaelis).

# Asks to confirm using the model instead of coordinates
if (interactive()) {
  str(
    propagate_errors(biomass, n = 50, height_model = model)
  )
}