Data from SCBI.

scbi_tree1 and scbi_species are complete tree and species tables from SCBI.

scbi_stem_shrub_tiny and scbi_stem_tree_tiny are small example datasets containing exclusively shrubs and trees.

scbi_tree1

scbi_species

scbi_stem_shrub_tiny

scbi_stem_tree_tiny

Format

Objects of class tbl_df (inherit from tbl, data.frame). See structure in Examples.

Source

http://bit.ly/scbi-forestgeo-data

Examples

str(scbi_tree1)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	40283 obs. of  20 variables:
#>  $ treeID   : int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ stemID   : int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ tag      : chr  "10079" "10168" "10567" "12165" ...
#>  $ StemTag  : chr  "1" "1" "1" "1" ...
#>  $ sp       : chr  "libe" "libe" "libe" "nysy" ...
#>  $ quadrat  : chr  "0104" "0103" "0110" "0122" ...
#>  $ gx       : num  3.7 17.3 9 14.2 9.4 ...
#>  $ gy       : num  73 58.9 197.1 428.5 436.4 ...
#>  $ DBHID    : int  1 3 5 7 9 13 15 17 19 22 ...
#>  $ CensusID : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ dbh      : num  27.9 23.7 22.2 135 87 ...
#>  $ pom      : chr  "1.3" "1.3" "1.3" "1.3" ...
#>  $ hom      : num  1.3 1.3 1.3 1.3 1.3 ...
#>  $ ExactDate: chr  "2008-10-21" "2008-10-21" "2008-10-22" "2008-09-08" ...
#>  $ DFstatus : chr  "alive" "alive" "alive" "alive" ...
#>  $ codes    : chr  "M" "M,X" "M" "M" ...
#>  $ nostems  : num  2 2 2 2 4 2 2 2 3 2 ...
#>  $ date     : num  17826 17826 17827 17783 17783 ...
#>  $ status   : chr  "A" "A" "A" "A" ...
#>  $ agb      : num  0.001586 0.000974 0.000988 0.123603 0.051184 ...

str(scbi_stem_shrub_tiny)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	34 obs. of  20 variables:
#>  $ treeID   : num  1111 1111 1111 1111 1111 ...
#>  $ stemID   : num  1111 32221 36962 38540 39225 ...
#>  $ tag      : num  10133 10133 10133 10133 10133 ...
#>  $ StemTag  : num  1 2 3 4 5 6 1 10 11 12 ...
#>  $ sp       : chr  "libe" "libe" "libe" "libe" ...
#>  $ quadrat  : chr  "0102" "0102" "0102" "0102" ...
#>  $ gx       : num  15.2 15.2 15.2 15.2 15.2 ...
#>  $ gy       : num  26.1 26.1 26.1 26.1 26.1 ...
#>  $ DBHID    : num  3295 3296 3297 3298 3299 ...
#>  $ CensusID : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ dbh      : num  50.9 18.5 17.2 17.2 17.1 ...
#>  $ pom      : num  1.3 1.3 1.3 1.3 1.3 1.3 1.3 1.3 1.3 1.3 ...
#>  $ hom      : num  1.3 1.3 1.3 1.3 1.3 ...
#>  $ ExactDate: Date, format: "2008-10-21" "2008-10-21" ...
#>  $ DFstatus : chr  "alive" "alive" "alive" "alive" ...
#>  $ codes    : chr  "M" "S" "S" "S" ...
#>  $ countPOM : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ date     : num  17826 17826 17826 17826 17826 ...
#>  $ status   : chr  "A" "A" "A" "A" ...
#>  $ agb      : num  0.00581 0.000463 0.00039 0.00039 0.000385 ...

str(scbi_stem_tree_tiny)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	12 obs. of  20 variables:
#>  $ treeID   : num  509 509 12427 12427 12427 ...
#>  $ stemID   : num  509 31693 12427 33757 37528 ...
#>  $ tag      : num  121455 121455 100776 100776 100776 ...
#>  $ StemTag  : num  1 2 1 2 3 4 1 2 1 2 ...
#>  $ sp       : chr  "fram" "fram" "caca" "caca" ...
#>  $ quadrat  : chr  "1216" "1216" "1005" "1005" ...
#>  $ gx       : num  236 236 184 184 184 ...
#>  $ gy       : num  303.6 303.6 99.7 99.7 99.7 ...
#>  $ DBHID    : num  1656 1657 17376 17377 17378 ...
#>  $ CensusID : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ dbh      : num  229.6 215.4 125.2 18.6 14.6 ...
#>  $ pom      : num  1.3 1.3 1.3 1.3 1.3 1.3 1.3 1.3 1.3 1.3 ...
#>  $ hom      : num  1.3 1.3 1.3 1.3 1.3 ...
#>  $ ExactDate: Date, format: "2008-11-14" "2008-11-14" ...
#>  $ DFstatus : chr  "alive" "alive" "alive" "alive" ...
#>  $ codes    : chr  "M" "S" "M" "S" ...
#>  $ countPOM : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ date     : num  17850 17850 17842 17842 17842 ...
#>  $ status   : chr  "A" "A" "A" "A" ...
#>  $ agb      : num  0.331133 0.280111 0.070354 0.000528 0.0003 ...

str(scbi_species)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	73 obs. of  10 variables:
#>  $ sp       : chr  "acne" "acpl" "acru" "acsp" ...
#>  $ Latin    : chr  "Acer negundo" "Acer platanoides" "Acer rubrum" "Acer sp" ...
#>  $ Genus    : chr  "Acer" "Acer" "Acer" "Acer" ...
#>  $ Species  : chr  "negundo" "platanoides" "rubrum" "sp" ...
#>  $ Family   : chr  "Sapindaceae" "Sapindaceae" "Sapindaceae" "Sapindaceae" ...
#>  $ SpeciesID: int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ Authority: chr  "" "" "" "" ...
#>  $ IDLevel  : chr  "species" "species" "species" "Multiple" ...
#>  $ syn      : logi  NA NA NA NA NA NA ...
#>  $ subsp    : logi  NA NA NA NA NA NA ...