Valid units as per measurements::conv_unit_options.

This simply wraps measurements::conv_unit_options to provide a more evocative name, closer to the problem domain. You may use this function to know how to name units in a way that ensures that units conversion will be converted correctly.

valid_units()

Value

A list

See also

Examples

valid_units()

#> $acceleration
#>  [1] "mm_per_sec2"   "cm_per_sec2"   "m_per_sec2"    "km_per_sec2"  
#>  [5] "grav"          "inch_per_sec2" "ft_per_sec2"   "mi_per_sec2"  
#>  [9] "kph_per_sec"   "mph_per_sec"  
#> 
#> $angle
#> [1] "degree" "radian" "grad"   "arcmin" "arcsec" "turn"  
#> 
#> $area
#>  [1] "nm2"      "um2"      "mm2"      "cm2"      "m2"       "hectare" 
#>  [7] "km2"      "inch2"    "in2"      "ft2"      "yd2"      "acre"    
#> [13] "mi2"      "naut_mi2"
#> 
#> $coordinate
#> [1] "dec_deg"     "deg_dec_min" "deg_min_sec" "dms"         "DMS"        
#> 
#> $count
#> [1] "fmol" "pmol" "nmol" "umol" "mmol" "mol" 
#> 
#> $duration
#>  [1] "nsec"      "usec"      "msec"      "sec"       "min"       "hr"       
#>  [7] "hour"      "day"       "wk"        "week"      "mon"       "yr"       
#> [13] "year"      "dec"       "decade"    "cen"       "century"   "mil"      
#> [19] "millenium" "Ma"       
#> 
#> $energy
#> [1] "J"    "kJ"   "erg"  "cal"  "Cal"  "Wsec" "kWh"  "MWh"  "BTU" 
#> 
#> $file_size
#>  [1] "byte" "KB"   "kB"   "MB"   "GB"   "TB"   "PB"   "bit"  "Kbit" "kbit"
#> [11] "Mbit" "Gbit" "Tbit" "Pbit"
#> 
#> $flow
#>  [1] "ml_per_sec"  "ml_per_min"  "ml_per_hr"   "l_per_sec"   "l_per_min"  
#>  [6] "LPM"         "l_per_hr"    "m3_per_sec"  "m3_per_min"  "m3_per_hr"  
#> [11] "Sv"          "gal_per_sec" "gal_per_min" "GPM"         "gal_per_hr" 
#> [16] "ft3_per_sec" "ft3_per_min" "ft3_per_hr" 
#> 
#> $length
#>  [1] "angstrom"   "nm"         "um"         "mm"         "cm"        
#>  [6] "dm"         "m"          "km"         "inch"       "ft"        
#> [11] "foot"       "feet"       "yd"         "yard"       "fathom"    
#> [16] "mi"         "mile"       "naut_mi"    "au"         "light_yr"  
#> [21] "light_year" "parsec"     "point"     
#> 
#> $mass
#>  [1] "Da"         "fg"         "pg"         "ng"         "ug"        
#>  [6] "mg"         "g"          "kg"         "Mg"         "Gg"        
#> [11] "Tg"         "Pg"         "carat"      "metric_ton" "oz"        
#> [16] "lbs"        "short_ton"  "long_ton"   "stone"     
#> 
#> $power
#>  [1] "uW"          "mW"          "W"           "kW"          "MW"         
#>  [6] "GW"          "erg_per_sec" "cal_per_sec" "cal_per_hr"  "Cal_per_sec"
#> [11] "Cal_per_hr"  "BTU_per_sec" "BTU_per_hr"  "hp"         
#> 
#> $pressure
#>  [1] "uatm"  "atm"   "Pa"    "hPa"   "kPa"   "torr"  "mmHg"  "inHg"  "cmH2O"
#> [10] "inH2O" "mbar"  "bar"   "dbar"  "psi"   "PSI"  
#> 
#> $speed
#>  [1] "mm_per_sec"   "cm_per_sec"   "m_per_sec"    "km_per_sec"   "inch_per_sec"
#>  [6] "ft_per_sec"   "kph"          "km_per_hr"    "mph"          "mi_per_hr"   
#> [11] "km_per_day"   "mi_per_day"   "knot"         "mach"         "light"       
#> 
#> $temperature
#> [1] "C" "F" "K" "R"
#> 
#> $volume
#>  [1] "ul"        "ml"        "dl"        "l"         "cm3"       "dm3"      
#>  [7] "m3"        "km3"       "us_tsp"    "us_tbsp"   "us_oz"     "us_cup"   
#> [13] "us_pint"   "us_quart"  "us_gal"    "inch3"     "ft3"       "mi3"      
#> [19] "imp_tsp"   "imp_tbsp"  "imp_oz"    "imp_cup"   "imp_pint"  "imp_quart"
#> [25] "imp_gal"  
#>