Add allometric equations to a census dataset.

This function adds columns to uniquely identify each row of a census dataset, and each equation from allodb -- so you can look it up on the database.

add_equations(census, equations)

Arguments

census
A ForestGEO-like census dataframe.

equations
An equations dataframe with unique rowids.

Value

A dataframe with all columns from census plus additional columns rowid, eqn equation_id.

See also

Examples

census <- fgeo.biomass::scbi_tree1
species <- fgeo.biomass::scbi_species

single_best <- census %>%
  add_species(species, site = "scbi") %>%
  allo_find() %>%
  allo_order() %>%
  fixme_pick_one_row_by_rowid()

add_equations(census, single_best)

#> Joining, by = "rowid"

#> # A tibble: 40,283 x 23
#>    rowid treeID stemID tag   StemTag sp    quadrat    gx    gy DBHID CensusID
#>    <int>  <int>  <int> <chr> <chr>   <chr> <chr>   <dbl> <dbl> <int>    <int>
#>  1     1      1      1 10079 1       libe  0104     3.70  73       1        1
#>  2     2      2      2 10168 1       libe  0103    17.3   58.9     3        1
#>  3     3      3      3 10567 1       libe  0110     9    197.      5        1
#>  4     4      4      4 12165 1       nysy  0122    14.2  428.      7        1
#>  5     5      5      5 12190 1       havi  0122     9.40 436.      9        1
#>  6     6      6      6 12192 1       havi  0122     1.30 434      13        1
#>  7     7      7      7 12212 1       unk   0123    17.8  447.     15        1
#>  8     8      8      8 12261 1       libe  0125    18    484.     17        1
#>  9     9      9      9 12456 1       vipr  0130    18    598.     19        1
#> 10    10     10     10 12551 1       astr  0132     5.60 628.     22        1
#> # … with 40,273 more rows, and 12 more variables: dbh <dbl>, pom <chr>,
#> #   hom <dbl>, ExactDate <chr>, DFstatus <chr>, codes <chr>, nostems <dbl>,
#> #   date <dbl>, status <chr>, agb <dbl>, eqn <chr>, equation_id <chr>