Get default equations of each type.

Get default equations of each type.

allo_find(dbh_species, custom_eqn = NULL)

Arguments

dbh_species
A dataframe as those created with add_species().

custom_eqn
A dataframe of class "eqn".

Value

A nested dataframe with each row containing the data of an equation type.

See also

Examples

census <- dplyr::sample_n(fgeo.biomass::scbi_tree1, 30)
species <- fgeo.biomass::scbi_species
census_species <- add_species(
  census, species,
  site = "scbi"
)

#> `sp` now stores Latin species names

allo_find(census_species)

#> Joining, by = c("sp", "site")

#> Warning: The input and output datasets have different number of rows:
#> * Input: 30.
#> * Output: 24.

#> Converting `dbh` based on `dbh_unit`.

#> # A tibble: 24 x 11
#>    eqn_type rowid site  sp      dbh equation_id eqn   eqn_source
#>    <chr>    <int> <chr> <chr> <dbl> <chr>       <chr> <chr>     
#>  1 species      5 scbi  nyss…  54.7 8da09d      1.54… default   
#>  2 species      7 scbi  corn…  16.7 f3e023      3.08… default   
#>  3 species     10 scbi  quer…  31.1 07dba7      1.56… default   
#>  4 species     11 scbi  robi… 102.  333c34      4.13… default   
#>  5 species     11 scbi  robi… 102.  e9d686      1.04… default   
#>  6 species     14 scbi  liri…  76.6 34fe5a      1.02… default   
#>  7 species     20 scbi  frax…  24.1 973c05      2.36… default   
#>  8 species     21 scbi  quer… 122.  07dba7      1.56… default   
#>  9 genus        9 scbi  cary… 127.  9c4cc9      10^(… default   
#> 10 genus       12 scbi  cary…  60.8 9c4cc9      10^(… default   
#> # … with 14 more rows, and 3 more variables: anatomic_relevance <chr>,
#> #   dbh_unit <chr>, bms_unit <chr>

# PROVIDE CUSTOM EQUAITONS ----------------------------------------------
# Checks that the structure of your data isn't terriby wrong
# BAD
try(as_eqn("really bad data"))

#> Error in validate_eqn(data) : is.data.frame(data) is not TRUE

try(as_eqn(data.frame(1)))

#> Error : Ensure your data set has these variables:
#> equation_id, site, sp, eqn, eqn_type, anatomic_relevance, dbh_unit, bms_unit

# GOOD
your_equations <- tibble::tibble(
  equation_id = c("000001"),
  site = c("scbi"),
  sp = c("tilia americana"),
  # Watning: Fake!
  eqn = c("exp(-2.48 + 2.4835 * log(dbh))"),
  eqn_type = c("mixed_hardwood"),
  anatomic_relevance = c("total aboveground biomass"),
 dbh_unit = "cm",
 bms_unit = "g"
)

class(as_eqn(your_equations))

#> [1] "eqn"        "tbl_df"     "tbl"        "data.frame"

allo_find(census_species, custom_eqn = as_eqn(your_equations))

#> Joining, by = c("sp", "site")

#> Warning: The input and output datasets have different number of rows:
#> * Input: 30.
#> * Output: 0.

#> Converting `dbh` based on `dbh_unit`.

#> Warning: Unknown or uninitialised column: 'dbh'.

#> Warning: Unknown or uninitialised column: 'dbh'.

#> # A tibble: 0 x 10
#> # … with 10 variables: eqn_type <chr>, rowid <int>, site <chr>, sp <chr>,
#> #   equation_id <chr>, eqn <chr>, eqn_source <chr>, anatomic_relevance <chr>,
#> #   dbh_unit <chr>, bms_unit <chr>

census_species %>%
allo_find(custom_eqn = as_eqn(your_equations))

#> Joining, by = c("sp", "site")

#> Warning: The input and output datasets have different number of rows:
#> * Input: 30.
#> * Output: 0.

#> Converting `dbh` based on `dbh_unit`.

#> Warning: Unknown or uninitialised column: 'dbh'.

#> Warning: Unknown or uninitialised column: 'dbh'.

#> # A tibble: 0 x 10
#> # … with 10 variables: eqn_type <chr>, rowid <int>, site <chr>, sp <chr>,
#> #   equation_id <chr>, eqn <chr>, eqn_source <chr>, anatomic_relevance <chr>,
#> #   dbh_unit <chr>, bms_unit <chr>