Handle multiple rowids: find duplicates and pick a single row.

Handle multiple rowids: find duplicates and pick a single row.

fixme_find_duplicated_rowid(data)

fixme_drop_duplicated_rowid(data)

Arguments

data
A dataframe as those created with allo_find().

Value

fixme_find_duplicated_rowid() returns a dataframe with the rows for which duplicated rowid values are found.

fixme_drop_duplicated_rowid() returns a dataframe with a single row per rowid value.

See also

Examples

library(dplyr)

best <- fgeo.biomass::scbi_tree1 %>%
  sample_n(500) %>%
  add_species(fgeo.biomass::scbi_species, "scbi") %>%
  allo_find()

#> Adding `site`.

#> Overwriting `sp`; it now stores Latin species names.

#> Adding `rowid`.

#> Assuming `dbh` data in [mm].

#> Joining, by = c("sp", "site")

#> Converting `dbh` based on `dbh_unit`.

#> Warning: Can't convert all units (inserting 18 missing values):
#> the 'to' argument is not an acceptable unit.

best %>%
  fixme_find_duplicated_rowid()

#> # A tibble: 48 x 30
#>    rowid treeID stemID tag   StemTag sp    quadrat    gx    gy DBHID CensusID
#>    <int>  <int>  <int> <chr> <chr>   <chr> <chr>   <dbl> <dbl> <int>    <int>
#>  1     2   9374   9374 80292 1       nyss… 0806     149.  111. 13737        1
#>  2     2   9374   9374 80292 1       nyss… 0806     149.  111. 13737        1
#>  3     2   9374   9374 80292 1       nyss… 0806     149.  111. 13737        1
#>  4     6   7758   7758 62180 1       nyss… 0621     109   411. 11703        1
#>  5     6   7758   7758 62180 1       nyss… 0621     109   411. 11703        1
#>  6     6   7758   7758 62180 1       nyss… 0621     109   411. 11703        1
#>  7    20  13102  13102 1021… 1       nyss… 1020     196   391. 18193        1
#>  8    20  13102  13102 1021… 1       nyss… 1020     196   391. 18193        1
#>  9    20  13102  13102 1021… 1       nyss… 1020     196   391. 18193        1
#> 10    50   7636   7636 62058 1       amel… 0618     104.  346. 11572        1
#> # … with 38 more rows, and 19 more variables: dbh <dbl>, pom <chr>, hom <dbl>,
#> #   ExactDate <chr>, DFstatus <chr>, codes <chr>, nostems <dbl>, date <dbl>,
#> #   status <chr>, agb <dbl>, site <chr>, equation_id <chr>, eqn <chr>,
#> #   eqn_source <chr>, eqn_type <chr>, anatomic_relevance <chr>, dbh_unit <chr>,
#> #   bms_unit <chr>, n <int>

best %>%
  fixme_drop_duplicated_rowid()

#> Warning: Dropping 17 rows with duplicated `rowid` values.

#> # A tibble: 483 x 29
#>    rowid treeID stemID tag   StemTag sp    quadrat    gx     gy DBHID CensusID
#>    <int>  <int>  <int> <chr> <chr>   <chr> <chr>   <dbl>  <dbl> <int>    <int>
#>  1     1  11900  11900 1002… 1       lind… 1001    181.   10.1  16691        1
#>  2     3  13111  13111 1021… 1       cary… 1020    192.  397.   18202        1
#>  3     4   4043   4043 30412 1       asim… 0310     58.4 194.    7152        1
#>  4     5  13302  13302 1023… 1       liri… 1025    183.  493.   18397        1
#>  5     7  10830  10830 90585 1       cary… 0914    178   280.   15446        1
#>  6     8  35586     NA 1833… <NA>    lind… 1805    354    80.6     NA       NA
#>  7     9  27010  27010 1820… 1       asim… 1821    354.  409.   35243        1
#>  8    10   2994   2994 20811 1       cary… 0215     31   295.    5899        1
#>  9    11  27600  27600 1827… 1       frax… 1832    350   624.   35912        1
#> 10    12  24197  24197 1702… 1       lind… 1701    336.    3.80 31805        1
#> # … with 473 more rows, and 18 more variables: dbh <dbl>, pom <chr>, hom <dbl>,
#> #   ExactDate <chr>, DFstatus <chr>, codes <chr>, nostems <dbl>, date <dbl>,
#> #   status <chr>, agb <dbl>, site <chr>, equation_id <chr>, eqn <chr>,
#> #   eqn_source <chr>, eqn_type <chr>, anatomic_relevance <chr>, dbh_unit <chr>,
#> #   bms_unit <chr>

# Should return 0-rows
best %>%
  fixme_drop_duplicated_rowid() %>%
  fixme_find_duplicated_rowid()

#> Warning: Dropping 17 rows with duplicated `rowid` values.

#> # A tibble: 0 x 30
#> # … with 30 variables: rowid <int>, treeID <int>, stemID <int>, tag <chr>,
#> #   StemTag <chr>, sp <chr>, quadrat <chr>, gx <dbl>, gy <dbl>, DBHID <int>,
#> #   CensusID <int>, dbh <dbl>, pom <chr>, hom <dbl>, ExactDate <chr>,
#> #   DFstatus <chr>, codes <chr>, nostems <dbl>, date <dbl>, status <chr>,
#> #   agb <dbl>, site <chr>, equation_id <chr>, eqn <chr>, eqn_source <chr>,
#> #   eqn_type <chr>, anatomic_relevance <chr>, dbh_unit <chr>, bms_unit <chr>,
#> #   n <int>