Handle multiple rowids: find duplicates and pick a single row.

Handle multiple rowids: find duplicates and pick a single row.

fixme_find_duplicated_rowid(data)

fixme_drop_duplicated_rowid(data)

Arguments

data
A dataframe as those created with allo_order().

Value

fixme_find_duplicated_rowid() returns a dataframe with the rows for which duplicated rowid values are found.

fixme_drop_duplicated_rowid() returns a dataframe with a single row per rowid value.

See also

Examples

best <- fgeo.biomass::scbi_tree1 %>%
  add_species(fgeo.biomass::scbi_species, "scbi") %>%
  allo_find() %>%
  allo_order()

best %>%
  fixme_find_duplicated_rowid()

#> # A tibble: 1,809 x 10
#>    eqn_type rowid site  sp      dbh equation_id eqn   eqn_source
#>    <chr>    <int> <chr> <chr> <dbl> <chr>       <chr> <chr>     
#>  1 genus      106 scbi  amel…  13.8 c59e03      exp(… default   
#>  2 genus      106 scbi  amel…  13.8 96c0af      10^(… default   
#>  3 genus      106 scbi  amel…  13.8 529234      10^(… default   
#>  4 genus      125 scbi  amel…  36.1 c59e03      exp(… default   
#>  5 genus      125 scbi  amel…  36.1 96c0af      10^(… default   
#>  6 genus      125 scbi  amel…  36.1 529234      10^(… default   
#>  7 genus      131 scbi  amel…  79.1 c59e03      exp(… default   
#>  8 genus      131 scbi  amel…  79.1 96c0af      10^(… default   
#>  9 genus      131 scbi  amel…  79.1 529234      10^(… default   
#> 10 genus      132 scbi  amel…  24   c59e03      exp(… default   
#> # … with 1,799 more rows, and 2 more variables: anatomic_relevance <chr>,
#> #   n <int>

best %>%
  fixme_drop_duplicated_rowid()

#> Warning: Dropping 783 rows with duplicated `rowid` values.

#> # A tibble: 28,420 x 9
#>    eqn_type rowid site  sp      dbh equation_id eqn   eqn_source
#>    <chr>    <int> <chr> <chr> <dbl> <chr>       <chr> <chr>     
#>  1 family       1 scbi  lind…  27.9 f08fff      exp(… default   
#>  2 family       2 scbi  lind…  23.7 f08fff      exp(… default   
#>  3 family       3 scbi  lind…  22.2 f08fff      exp(… default   
#>  4 family       8 scbi  lind…  51.4 f08fff      exp(… default   
#>  5 family      13 scbi  lind…  15.4 f08fff      exp(… default   
#>  6 family      14 scbi  lind…  14.8 f08fff      exp(… default   
#>  7 family      15 scbi  lind…  15.5 f08fff      exp(… default   
#>  8 family      16 scbi  lind…  17.4 f08fff      exp(… default   
#>  9 family      17 scbi  lind…  68.2 f08fff      exp(… default   
#> 10 family      18 scbi  lind…  19.3 f08fff      exp(… default   
#> # … with 28,410 more rows, and 1 more variable: anatomic_relevance <chr>

best %>%
  fixme_drop_duplicated_rowid()

#> Warning: Dropping 783 rows with duplicated `rowid` values.

#> # A tibble: 28,420 x 9
#>    eqn_type rowid site  sp      dbh equation_id eqn   eqn_source
#>    <chr>    <int> <chr> <chr> <dbl> <chr>       <chr> <chr>     
#>  1 family       1 scbi  lind…  27.9 f08fff      exp(… default   
#>  2 family       2 scbi  lind…  23.7 f08fff      exp(… default   
#>  3 family       3 scbi  lind…  22.2 f08fff      exp(… default   
#>  4 family       8 scbi  lind…  51.4 f08fff      exp(… default   
#>  5 family      13 scbi  lind…  15.4 f08fff      exp(… default   
#>  6 family      14 scbi  lind…  14.8 f08fff      exp(… default   
#>  7 family      15 scbi  lind…  15.5 f08fff      exp(… default   
#>  8 family      16 scbi  lind…  17.4 f08fff      exp(… default   
#>  9 family      17 scbi  lind…  68.2 f08fff      exp(… default   
#> 10 family      18 scbi  lind…  19.3 f08fff      exp(… default   
#> # … with 28,410 more rows, and 1 more variable: anatomic_relevance <chr>

# Should return 0-rows
best %>%
  fixme_drop_duplicated_rowid() %>%
  fixme_find_duplicated_rowid()

#> Warning: Dropping 783 rows with duplicated `rowid` values.

#> # A tibble: 0 x 10
#> # … with 10 variables: eqn_type <chr>, rowid <int>, site <chr>, sp <chr>,
#> #   dbh <dbl>, equation_id <chr>, eqn <chr>, eqn_source <chr>,
#> #   anatomic_relevance <chr>, n <int>